requirements.txt
streamlit
openai
python-dotenv
streamlit-webrtc
numpy
soundfile
av
pandas
google-cloud-language
google-auth